Revisions to reward presentation
Add the "Fix" button to go back to the labeling screen. (It might be frustrated to load the screen again but still takes a period of time. I can optimize this in the future, but for now it's the fastest way to implement this feature.)
The program will load reward.csv in assets before it tries to load reward.csv in the external storage. So if the former csv file has the line "1, 1.html, http://xxx" and the latter one has "1, 11.html, http://xxx", the line in the former one will be replaced by "1, 11.html, http://xxx" in the memory because the first column is used as the key to html and link.

The Reward folder will be copied from assets to the external storage (uscteens/data/Reward) in the following cases: 1. The app's version changes and the user runs the new version of this app for the first time. 2. The Reward folder in the external storage is nonexistent, deleted or renamed.